refactoring for date extraction (use interfaces, generics, other style improvements)
git-svn-id: https://141.76.40.86/svn-students/iircommon/toolkit/trunk@2876 5654db93-19e5-4567-beb8-64e8c25c7c45

Former-commit-id: d43ebfba1a3357f5b940423926f4fb4e33a66bf9

+++ src/main/java/ws/palladian/extraction/date/dates/DateExactnessComparator.java
@@ -13,6 +13,6 @@ public class DateExactnessComparator implements Comparator<ExtractedDate> {
 	 */
 	@Override
 	public int compare(ExtractedDate ed1, ExtractedDate ed2) {
-		return (int) (ed2.getExactness() - ed1.getExactness());
+		return (ed2.getExactness() - ed1.getExactness());
 	}
 }
